Renamed CrosshairNarrator to WorldNarrator
