+++ common/src/main/java/org/mcaccess/minecraftaccess/Config.java
@@ -7,7 +7,7 @@ import me.shedaniel.autoconfig.ConfigData;
 import me.shedaniel.autoconfig.annotation.ConfigEntry;
 import net.minecraft.resources.ResourceLocation;
 
-import org.mcaccess.minecraftaccess.api.CrosshairNarrator;
+import org.mcaccess.minecraftaccess.api.WorldNarrator;
 import org.mcaccess.minecraftaccess.api.Status;
 import org.mcaccess.minecraftaccess.utils.config.ConfigExtension;
 
@@ -230,7 +230,7 @@ public final class Config implements ConfigData {
 
     public static final class NarrateCrosshair {
         public boolean enabled = true;
-        @ConfigExtension.Registry(registry = CrosshairNarrator.class, i18n = "narrator")
+        @ConfigExtension.Registry(registry = WorldNarrator.class, i18n = "narrator")
         public ResourceLocation narrator = ResourceLocation.fromNamespaceAndPath(MainClass.MOD_ID, Platform.isModLoaded("jade") ? "jade" : "minecraft_access");
         public boolean narrateBlockFace = true;
         public boolean disableNarratingConsecutiveBlocks = false;
