+++ common/src/main/java/org/mcaccess/minecraftaccess/addon/accessmenu/NarrateTarget.java
@@ -8,7 +8,7 @@ import net.minecraft.world.phys.HitResult;
 import org.mcaccess.minecraftaccess.Config;
 import org.mcaccess.minecraftaccess.MainClass;
 import org.mcaccess.minecraftaccess.api.AccessMenuFunction;
-import org.mcaccess.minecraftaccess.api.CrosshairNarrator;
+import org.mcaccess.minecraftaccess.api.WorldNarrator;
 import org.mcaccess.minecraftaccess.utils.NarrationUtils;
 import org.mcaccess.minecraftaccess.utils.PlayerUtils;
 
@@ -23,7 +23,7 @@ public class NarrateTarget implements AccessMenuFunction {
                 BlockHitResult blockHit = (BlockHitResult) hit;
                 BlockPos blockPos = blockHit.getBlockPos();
                 String narration = new StringBuilder()
-                        .append(MainClass.registry(CrosshairNarrator.class).get(Config.getInstance().narrateCrosshair.narrator).narrate(blockPos))
+                        .append(MainClass.registry(WorldNarrator.class).get(Config.getInstance().narrateCrosshair.narrator).narrate(blockPos))
                         .append(I18n.get("minecraft_access.other.words_connection"))
                         .append(NarrationUtils.narrateRelativePositionOfPlayerAnd(blockPos))
                         .toString();
