+++ common/src/main/java/org/mcaccess/minecraftaccess/api/AddonRegistry.java
@@ -65,16 +65,16 @@ public final class AddonRegistry {
     }
 
     /**
-     * Registers a {@link CrosshairNarrator}.
+     * Registers a {@link WorldNarrator}.
      *
-     * @param identifier A unique identifier for the {@link CrosshairNarrator}. This is automatically namespaced by mod ID.
-     * @param crosshairNarrator The {@link CrosshairNarrator} to register.
-     * @throws IllegalArgumentException If a {@link CrosshairNarrator} is already registered with the provided identifier and the same mod ID.
+     * @param identifier A unique identifier for the {@link WorldNarrator}. This is automatically namespaced by mod ID.
+     * @param worldNarrator The {@link WorldNarrator} to register.
+     * @throws IllegalArgumentException If a {@link WorldNarrator} is already registered with the provided identifier and the same mod ID.
      * @throws UnsupportedOperationException Registries are already frozen.
-     * @see CrosshairNarrator
+     * @see WorldNarrator
      * @since 1.12.0
      */
-    public void register(@NotNull String identifier, @NotNull CrosshairNarrator crosshairNarrator) {
-        MainClass.register(CrosshairNarrator.class, ResourceLocation.fromNamespaceAndPath(modid, identifier), crosshairNarrator);
+    public void register(@NotNull String identifier, @NotNull WorldNarrator worldNarrator) {
+        MainClass.register(WorldNarrator.class, ResourceLocation.fromNamespaceAndPath(modid, identifier), worldNarrator);
     }
 }
