+++ common/src/main/java/org/mcaccess/minecraftaccess/features/point_of_interest/LockingHandler.java
@@ -3,7 +3,6 @@ package org.mcaccess.minecraftaccess.features.point_of_interest;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 
 import lombok.extern.slf4j.Slf4j;
 import net.minecraft.client.Minecraft;
@@ -27,7 +26,7 @@ import net.minecraft.world.phys.Vec3;
 
 import org.mcaccess.minecraftaccess.Config;
 import org.mcaccess.minecraftaccess.MainClass;
-import org.mcaccess.minecraftaccess.api.CrosshairNarrator;
+import org.mcaccess.minecraftaccess.api.WorldNarrator;
 import org.mcaccess.minecraftaccess.utils.KeyBindingsHandler;
 import org.mcaccess.minecraftaccess.utils.NarrationUtils;
 import org.mcaccess.minecraftaccess.utils.PlayerUtils;
@@ -290,7 +289,7 @@ public class LockingHandler {
         lockedOnEntity = entity;
 
         StringBuilder narration = new StringBuilder(
-                MainClass.registry(CrosshairNarrator.class).get(Config.getInstance().narrateCrosshair.narrator).narrate(entity)
+                MainClass.registry(WorldNarrator.class).get(Config.getInstance().narrateCrosshair.narrator).narrate(entity)
         );
 
         if (Config.getInstance().poi.narrateDistance) {
@@ -318,7 +317,7 @@ public class LockingHandler {
         lockedOnBlockPos = new BlockPos3d(position, absolutePosition);
 
         StringBuilder blockDescription = new StringBuilder(
-                MainClass.registry(CrosshairNarrator.class).get(Config.getInstance().narrateCrosshair.narrator).narrate(lockedOnBlockPos)
+                MainClass.registry(WorldNarrator.class).get(Config.getInstance().narrateCrosshair.narrator).narrate(lockedOnBlockPos)
         );
         if (Config.getInstance().poi.narrateDistance) {
             blockDescription.append(' ')
